putty xserver xkeyboard-config libX11 pixman fontconfig mesa git update Sep 24 2019
xserver          commit 6036e845271800b2e268f5cba9283ce07a7fad3a
xkeyboard-config commit 75adcd543befa3c37c8e778b796176e525ba6091
libX11           commit edc7680ed5a03cedb5facf14693823455e12c29c
pixman           commit d60b0af5e32fed034bd163d244148947d3ed91a8
fontconfig       commit 65087ac7ce4cc5f2109967c1380b474955dcb590
mesa             commit da496d4e30b78a0a8e096453f3ffccffaf9c2f7e
putty            commit 8b87d80a8483c82f4e1479e45d26e103ab853154

+++ mesalib/src/gallium/auxiliary/driver_ddebug/dd_draw.c
@@ -57,7 +57,7 @@ dd_get_debug_filename_and_mkdir(char *buf, size_t buflen, bool verbose)
       fprintf(stderr, "dd: can't create a directory (%i)\n", errno);
 
    snprintf(buf, buflen, "%s/%s_%u_%08u", dir, proc_name, getpid(),
-            p_atomic_inc_return(&index) - 1);
+            (unsigned int)p_atomic_inc_return(&index) - 1);
 
    if (verbose)
       fprintf(stderr, "dd: dumping to file %s\n", buf);
